requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
openai>=1.0.0
flask>=3.0.0
gunicorn>=21.2.0



